move Queen class into main file
