+++ EightQueen.java
@@ -286,3 +286,22 @@ class Tile {
 		this.col = col;
 	}
 }
+
+class Queen {
+	int placedAtRow;
+	int placedAtColumn;
+	int X_Position; 
+	int Y_Position;
+	boolean inFocus;
+	boolean isDragging;
+	boolean placed;
+	boolean safe;
+	
+	Color color; //unsafe queen is red, safe queen is blue
+	
+	//constructor
+	public Queen(int x, int y) {
+		X_Position = x;
+		Y_Position = y;
+	}
+}
